python/requirements.txt
numpy>=1.19.1
gym==0.17.2
pyyaml>=5.3.1
imageio>=2.9.0
pygame>=2.0.0
matplotlib>=3.3.3

